+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Pre start game state
 - End game menu
 - Need For Type ascii art to the Start state
+- Real time stats display
 
 ### Improved
 - End game state